% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_report.R
\name{gen_report}
\alias{gen_report}
\title{Ploting your activity.}
\usage{
gen_report(path = getwd())
}
\arguments{
\item{path}{path a character vector of full path names; the default corresponds to the working directory, getwd().}
}
\value{
local data frame.
}
\description{
Ploting your activity.
}
\examples{
\dontrun{
gen_report()
}
}

