% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indicate_gh_contribution.R
\name{indicate_gh_contribution}
\alias{indicate_gh_contribution}
\title{Visualize GitHub Contribution.}
\usage{
indicate_gh_contribution(username)
}
\description{
Visualize GitHub Contribution.
}

