% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicate_yr_activity.R
\name{indicate_yr_activity}
\alias{indicate_yr_activity}
\title{Ploting your activity.}
\usage{
indicate_yr_activity(df)
}
\arguments{
\item{df}{data frame.}
}
\description{
Ploting your activity.
}

