% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/list_rfiles.R
\name{list_rfiles}
\alias{list_rfiles}
\title{List of R and Image files.}
\usage{
list_rfiles(path = getwd())
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory, getwd().}
}
\value{
The list r and image filenames.
}
\description{
List of R and Image files.
}
\note{
checked 141210
}

