% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivation.R
\name{motivation}
\alias{motivation}
\title{Show a Random Motivational Message}
\usage{
motivation()
}
\description{
This function prints a random motivational message to keep you going.
}
\examples{
motivation()
}
