% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/motivator.R
\docType{package}
\name{motivator}
\alias{motivator}
\alias{motivator-package}
\title{motivator: Keep and improve working motivation.}
\description{
Visualize your R working activity log.
}
\author{
Shinya Uryu
}

