% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinchoku_test.R
\name{shinchoku_test}
\alias{shinchoku_test}
\title{What about SHINCHOKU}
\usage{
shinchoku_test(goal = 1, num = 1)
}
\arguments{
\item{goal}{interger.}

\item{num}{integer. select yeah voice.}
}
\description{
What about SHINCHOKU
}
\examples{
\dontrun{
shinchoku_test(goal = 10, num = 3)
}
}

