% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicate_gh_contribution.R
\name{indicate_gh_contribution}
\alias{indicate_gh_contribution}
\title{Visualize GitHub Contribution.}
\usage{
indicate_gh_contribution(username = getOption("github.user"))
}
\arguments{
\item{username}{character. Display user name (GitHub account).}
}
\description{
Visualize GitHub Contribution.
}

